Generates utility values for products
We express the utility of an item in the range (0.01-1.0). 
We divide this range into 5 categories [(0.01,0.2), (0.21,0.4), (0.41,0.6), (0.61,0.8), (0.81,1.0)]. 
We also specify a certain percentage of items for each category which would typically be found in a retail store (to maintain fairness in random allocation)
Then, we randomly allocate each product to a category and further randomly allocate them a value within that range.